refactor: extract JSONL utils, session utils, and ESM-ify adapters
- claudeville/adapters/jsonl-utils.js: shared readLines + parseJsonLines
  (was duplicated verbatim in openclaw, copilot, codex, vscode)
- shared/session-utils.js: shared normalizeTokens (was duplicated in
  collector/index.ts and claudeville/adapters/index.ts)
- 4 adapters updated to import from jsonl-utils.js instead of defining
  their own readLines/parseJsonLines
- vscode.ts: remove unused toTimestamp() helper, inline its single use
- adapters/index.ts: convert to pure ESM imports
- collector/index.ts: use shared/session-utils.js
