+++ collector/index.ts
@@ -4,6 +4,7 @@ const { createFileWatchers } = require('../shared/watch-utils');
 const crypto = require('crypto');
 const { adapters, getAllSessions, getAllWatchPaths, getActiveProviders, getSessionDetailByProvider } = require('../claudeville/adapters');
 import { estimateCost } from '../shared/cost.js';
+import { normalizeTokens } from '../shared/session-utils.js';
 
 const HUB_URL = process.env.HUB_URL || 'http://localhost:3030';
 const HUB_AUTH_TOKEN = process.env.HUB_AUTH_TOKEN || 'dev-secret';
@@ -19,18 +20,12 @@ let sending = false;
 let lastSentHash = '';
 
 function normalizeSession(session, detail) {
-  const tokenUsage = detail?.tokenUsage || session.tokenUsage || null;
-  const tokens = tokenUsage
-    ? {
-        input: Number(tokenUsage.totalInput || 0),
-        output: Number(tokenUsage.totalOutput || 0),
-      }
-    : { input: 0, output: 0 };
+  const tokens = normalizeTokens(detail?.tokenUsage, session.tokens || null);
 
   return {
     ...session,
     tokens,
-    tokenUsage,
+    tokenUsage: detail?.tokenUsage || null,
     estimatedCost: estimateCost(session.model, tokens),
   };
 }
